Change type of Style::stroke_width to u32

+++ embedded-graphics/src/fonts/font_builder.rs
@@ -113,7 +113,7 @@ where
         self
     }
 
-    fn stroke_width(self, _width: u8) -> Self {
+    fn stroke_width(self, _width: u32) -> Self {
         // Noop
 
         self
